Use ff standard range 0-65535 for autocenter force

+++ hid-t150/hid-t150.h
@@ -35,7 +35,7 @@ struct t150
 	struct {
 		spinlock_t access_lock;
 
-		uint8_t autocenter_force;
+		uint16_t autocenter_force;
 		bool autocenter_enabled;
 		uint16_t range;
 		uint16_t gain;
